pyproject.toml
[project]
name = "keryx"
version = "0.2.0"
description = "Production-grade LinkedIn outreach orchestration pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.0.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langgraph>=0.4.0",
    "oauth2client>=4.1.3",
    "playwright>=1.50.0",
    "playwright-stealth>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
